+++ services/data-collector/src/kline_aggregator.rs
@@ -134,8 +134,8 @@ impl KlineAggregator {
 
     /// 更新或创建窗口
     async fn update_window(&self, quote: &StockQuote, period: KlinePeriod) -> Option<KlineData> {
-        // 从 i64 timestamp 转换为 DateTime<Utc>
-        let current_time = chrono::DateTime::from_timestamp(quote.timestamp, 0)
+        // 从 u64 timestamp 转换为 DateTime<Utc>
+        let current_time = chrono::DateTime::from_timestamp(quote.timestamp as i64, 0)
             .unwrap_or_else(|| chrono::Utc::now());
         let window_key = Self::make_window_key(&quote.code, period, &current_time);
